Service layeer

+++ src/service/impl/JobVacancyServiceImpl.java
@@ -2,6 +2,7 @@ package service.impl;
 
 import bean.EmployeeHR;
 import bean.JobVacancy;
+import controller.command.CommandName;
 import dao.JobVacancyDAO;
 import dao.exception.DAOException;
 import dao.factory.DAOFactory;
@@ -13,12 +14,20 @@ import service.exception.ServiceException;
 import java.util.ArrayList;
 import java.util.List;
 
+import static service.validation.ServiceValidation.*;
+
 /**
- * Created by irinaleibutina on 05.04.17.
+ * Class implements {@link src.dao.JobVacancyService}
  */
 public class JobVacancyServiceImpl implements JobVacancyService {
     private static final Logger logger = LogManager.getLogger(JobVacancyServiceImpl.class.getName());
 
+    /**
+     * Method performs a dao level call to get all vacancies
+     *
+     * @return list of instances of {@link JobVacancy}
+     * @throws ServiceException
+     */
     @Override
     public List<JobVacancy> getVacancies() throws ServiceException {
         DAOFactory factory = DAOFactory.getInstance();
@@ -27,16 +36,22 @@ public class JobVacancyServiceImpl implements JobVacancyService {
         try {
             jobVacancies = jobVacancyDAO.getVacancies();
         } catch (DAOException e) {
-            logger.error(e);
             throw new ServiceException(e);
         }
         return jobVacancies;
     }
 
+    /**
+     * Method checks the input parameters performs a dao level call
+     * to add job vacancy
+     *
+     * @param jobVacancy current job vacancy
+     * @throws ServiceException
+     */
     @Override
     public void addVacancy(JobVacancy jobVacancy) throws ServiceException {
 
-        if (!validateEmployeeInfo(jobVacancy)) {
+        if (!validateJobVacancy(jobVacancy)) {
             throw new ServiceException("information is not correct");
         }
 
@@ -45,11 +60,17 @@ public class JobVacancyServiceImpl implements JobVacancyService {
         try {
             jobVacancyDAO.addVacancy(jobVacancy);
         } catch (DAOException e) {
-            logger.error(e);
             throw new ServiceException(e);
         }
     }
 
+    /**
+     * Method checks the input parameters performs a dao level call
+     * to delete job vacancy
+     *
+     * @param jobVacancyId
+     * @throws ServiceException
+     */
     @Override
     public void deleteVacancy(String jobVacancyId) throws ServiceException {
         if (!validateParam(jobVacancyId)) {
@@ -60,15 +81,26 @@ public class JobVacancyServiceImpl implements JobVacancyService {
         JobVacancyDAO jobVacancyDAO = daoFactory.getJobVacancyDAO();
         try {
             int jobVacancyIntId = Integer.parseInt(jobVacancyId);
+            if (!validateId(jobVacancyIntId)) {
+                throw new ServiceException();
+            }
             jobVacancyDAO.deleteVacancy(jobVacancyIntId);
-        } catch (DAOException e) {
+        } catch (IllegalArgumentException | NullPointerException | DAOException e) {
             logger.error(e);
             throw new ServiceException(e);
         }
     }
 
+    /**
+     * Method checks the input parameters performs a dao level call
+     * to get job vacancy info
+     *
+     * @param jobVacancyId
+     * @return instance of {@link JobVacancy}
+     * @throws ServiceException
+     */
     @Override
-    public JobVacancy searchVacancy(String jobVacancyId) throws ServiceException {
+    public JobVacancy getVacancy(String jobVacancyId) throws ServiceException {
 
         if (!validateParam(jobVacancyId)) {
             logger.error("invalid jobVacancyId");
@@ -79,15 +111,48 @@ public class JobVacancyServiceImpl implements JobVacancyService {
         JobVacancyDAO jobVacancyDAO = daoFactory.getJobVacancyDAO();
         try {
             int jobVacancyIntId = Integer.parseInt(jobVacancyId);
-            return jobVacancyDAO.searchVacancy(jobVacancyIntId);
-        } catch (DAOException e) {
+
+            if (!validateId(jobVacancyIntId)) {
+                throw new ServiceException();
+            }
+            return jobVacancyDAO.getVacancy(jobVacancyIntId);
+        } catch (IllegalArgumentException | NullPointerException | DAOException e) {
             logger.error(e);
             throw new ServiceException(e);
         }
     }
 
+    /**
+     * Method checks the input parameters performs a dao level call
+     * to restore vacancy
+     *
+     * @param title title of job vacancy
+     * @throws ServiceException
+     */
     @Override
     public void restoreVacancy(String title) throws ServiceException {
+        if (!validateParam(title)) {
+            logger.error("invalid title");
+            throw new ServiceException("information is not correct");
+        }
+        DAOFactory daoFactory = DAOFactory.getInstance();
+        JobVacancyDAO jobVacancyDAO = daoFactory.getJobVacancyDAO();
+        try {
+            jobVacancyDAO.restoreVacancy(title);
+        } catch (DAOException e) {
+            throw new ServiceException(e);
+        }
+    }
+
+    /**
+     * Method checks the input parameters performs a dao level call
+     * to search vacancy
+     *
+     * @param title title of job vacancy
+     * @throws ServiceException
+     */
+    @Override
+    public JobVacancy searchVacancy(String title) throws ServiceException {
         if (!validateParam(title)) {
             logger.error("invalid title");
             throw new ServiceException("information is not correct");
@@ -96,32 +161,9 @@ public class JobVacancyServiceImpl implements JobVacancyService {
         DAOFactory daoFactory = DAOFactory.getInstance();
         JobVacancyDAO jobVacancyDAO = daoFactory.getJobVacancyDAO();
         try {
-            jobVacancyDAO.restoreVacancy(title);
+            return jobVacancyDAO.searchVacancy(title);
         } catch (DAOException e) {
-            logger.error(e);
             throw new ServiceException(e);
         }
-    }
-
-    private boolean validateEmployeeInfo(JobVacancy jobVacancy){
-        String title = jobVacancy.getJobTitle();
-        if (!validateParam(title)) {
-            logger.error("invalid title");
-           return false;
-        }
-
-        String description = jobVacancy.getDescription();
-        if (!validateParam(description)) {
-            logger.error("invalid description");
-            return false;
-        }
-
-        return true;
-    }
-    private boolean validateParam(String param) {
-        if ((param == null) || (param.isEmpty())) {
-            return false;
-        }
-        return true;
     }
 }
